refactor: extract outlier detection form helper components
